+++ src/main/java/frc/robot/RobotContainer.java
@@ -18,10 +18,10 @@ import frc.robot.autoRoutines.ShootCommandGroup;
 import frc.robot.autoRoutines.ShootWithTurnCommandGroup;
 import frc.robot.commands.AllInCommand;
 import frc.robot.commands.AllOutCommand;
-import frc.robot.commands.climbing.ClimbDownCommand;
-import frc.robot.commands.climbing.ClimbUpCommand;
+import frc.robot.commands.climbing.ClimberDownCommand;
+import frc.robot.commands.climbing.ClimberUpCommand;
 import frc.robot.commands.drivetrain.DriveForDistanceCommand;
-import frc.robot.commands.drivetrain.ResetEncodersCommand;
+import frc.robot.commands.drivetrain.ResetDriveEncodersCommand;
 import frc.robot.commands.drivetrain.StopDriveCommand;
 import frc.robot.commands.drivetrain.TurnNoPIDCommand;
 import frc.robot.commands.drivetrain.TurnToTargetPIDCommand;
